requirements.txt
asgiref==3.6.0
astroid==2.14.2
boto3==1.26.73
botocore==1.29.73
colorama==0.4.6
dill==0.3.6
Django==4.1.7
django-storages==1.13.2
gunicorn==20.1.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomlkit==0.11.6
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
